# api/docs/swagger.yaml
definitions:
  models.Film:
    properties:
      description:
        type: string
      film_id:
        type: integer
      fulltext:
        type: string
      language_id:
        type: integer
      last_update:
        type: string
      length:
        type: integer
      rating: {}
      release_year:
        type: integer
      rental_duration:
        type: integer
      rentalRate:
        type: number
      replacement_cost:
        type: number
      special_features:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /film:
    get:
      consumes:
      - application/json
      description: Get List Film
      operationId: get_list_film
      produces:
      - application/json
      responses:
        "200":
          description: GetFilmBody
          schema:
            items:
              $ref: '#/definitions/models.Film'
            type: array
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Get List Film
      tags:
      - Film
    post:
      consumes:
      - application/json
      description: Create Film
      operationId: create_film
      parameters:
      - description: CreateFilmRequestBody
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/models.Film'
      produces:
      - application/json
      responses:
        "201":
          description: GetFilmBody
          schema:
            $ref: '#/definitions/models.Film'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Create Film
      tags:
      - Film
  /film/:
    put:
      consumes:
      - application/json
      description: Update Film
      operationId: update_film
      parameters:
      - description: CreateFilmRequestBody
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/models.Film'
      produces:
      - application/json
      responses:
        "200":
          description: GetFilmsBody
          schema:
            $ref: '#/definitions/models.Film'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Update Film
      tags:
      - Film
  /film/{id}:
    delete:
      consumes:
      - application/json
      description: Delete By Id Film
      operationId: delete_by_id_film
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetFilmBody
          schema:
            $ref: '#/definitions/models.Film'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Delete By Id Film
      tags:
      - Film
    get:
      consumes:
      - application/json
      description: Get By Id Film
      operationId: get_by_id_film
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetFilmBody
          schema:
            $ref: '#/definitions/models.Film'
        "400":
          description: Invalid Argument
          schema:
            type: string
        "500":
          description: Server Error
          schema:
            type: string
      summary: Get By Id Film
      tags:
      - Film
swagger: "2.0"